Switched fully to DB
