+++ filerepo/webapp/domain/file/file.py
@@ -1,28 +1,19 @@
 from typing import Optional
+from sqlalchemy import Column, Float, Integer, String, LargeBinary
+from filerepo.webapp.repository.database import Base
 
-class File:
+class File(Base):
 
-    """File represents your collection of files as an entity."""
-
-    def __init__(
-        self,
-        id: str,
-        file_name: str,
-        file_size: float,
-        file_type: str,
-        file_hash: str,
-        file_content: Optional[bytes] = None,
-        file_creation_time: Optional[float] = None,
-        file_update_time: Optional[float] = None,
-    ):
-        self.id: str = id
-        self.file_name: str = file_name
-        self.file_size: float = file_size
-        self.file_type: str = file_type
-        self.file_hash: str = file_hash
-        self.file_content: Optional[bytes] = file_content
-        self.file_creation_time: Optional[float] = file_creation_time
-        self.file_update_time: Optional[float] = file_update_time
+    """File represents your collection of files as an entity. This class also represents the DB model Files"""
+    __tablename__ = "Files"
+    id: int = Column(Integer, primary_key=True, nullable=False)
+    file_name: str = Column(String, nullable=False)
+    file_size: float = Column(Float, nullable=False)
+    file_type: str = Column(String, nullable=False)
+    file_hash: str = Column(String, nullable=False)
+    file_content: Optional[bytes] = Column(LargeBinary,nullable=False)
+    file_creation_time: Optional[float] = Column(Float, nullable=False)
+    file_update_time: Optional[float] = Column(Float, nullable=False)
 
     def __eq__(self, o: object) -> bool:
         if isinstance(o, File):
