+++ filerepo/webapp/repository/file/file_repository.py
@@ -4,21 +4,21 @@ import hashlib
 import time
 
 from filerepo.webapp.repository.file.file_dto import FileDTO
-from filerepo.webapp.repository.file_system import FileSystem
+from sqlalchemy.orm.session import Session
 from filerepo.webapp.domain.file.file_repository import FileRepository
 from filerepo.webapp.domain.file.file import File
 from filerepo.webapp.schemas.DTO.file_upload_model import FileUploadModel
 
 
 class FileRepositoryImpl(FileRepository):
-    """FileRepositoryImpl implements CRUD operations related File entity using File_System."""
+    """FileRepositoryImpl implements CRUD operations related File entity using SQLAlchemy."""
 
-    def __init__(self, file_system: FileSystem):
-        self.file_system: FileSystem = file_system
+    def __init__(self, session: Session):
+        self.session: Session = session
 
-    def find_by_id(self, id: str) -> FileDTO:
+    def find_by_id(self, id: int) -> FileDTO:
         try:
-            file_dto = FileDTO.from_entity(self.file_system.read(id))
+            file_dto = FileDTO.from_entity(self.session.query(File).filter_by(id=id).one())
         except:
             raise
 
@@ -26,38 +26,41 @@ class FileRepositoryImpl(FileRepository):
 
     def find_all(self) -> List[FileDTO]:
         try:
-            files_list = self.file_system.list_files()
-            filesDTO_list:List [FileDTO] = []
-            for file in files_list:
-                filesDTO_list.append(FileDTO.from_entity(file))
+            file_list = self.session.query(File).all()
+            fileDTO_list: List[FileDTO] = []
+            for file in file_list:
+                fileDTO_list.append(FileDTO.from_entity(file))
         except:
             raise
 
-        return filesDTO_list
+        return fileDTO_list
 
     def create(self, file_uploaded: FileUploadModel) -> File:
         try:
-            id: str = shortuuid.uuid()
-            file_type: str = file_uploaded.file_type
             hash = hashlib.sha256(file_uploaded.file_content).hexdigest()
             current_time = time.time()
-            file = File(id,file_uploaded.file_name,len(file_uploaded.file_content),file_type,hash,file_uploaded.file_content,current_time,current_time)
-            self.file_system.write(file)
+            file = File(file_name=file_uploaded.file_name, file_size=len(file_uploaded.file_content),
+                        file_type=file_uploaded.file_type, file_hash=hash, file_content=file_uploaded.file_content,
+                        file_creation_time=current_time, file_update_time=current_time)
+            self.session.add(file)
+            self.session.commit()
             return file
         except:
             raise
 
-    def delete_by_id(self, id: str):
+    def delete_by_id(self, id: int):
         try:
-            self.file_system.delete(id)
+            uploadActivity = self.session.query(File).get(id)
+            self.session.delete(uploadActivity)
+            self.session.commit()
         except:
             raise
 
     def find_by_hash(self, hash):
         try:
-            files = self.find_all()
-            for file in files:
-                if file.file_hash == hash:
-                    return file.id
+            file = self.session.query(File).filter(File.file_hash == hash).one_or_none()
+            if (file == None):
+                return None
+            return file.id
         except:
             raise
